compiler/decompetition_disassembler/requirements.txt
capstone
diff_match_patch
intervaltree
pyelftools
pyyaml
